Refactor GroovycWrapper into two parts
- One part is the TreeParser and takes care of parsing symbols.
- The other part is the WorkspaceCompiler which takes care of compiling the code.
- The CompilerWrapperProvider is gone and now the CompilerWrapper is just part of the shared config.
- New packages names: ls.api, ls.groovy.compilation, ls.groovy.services

Closes #84
