+++ groovy-language-server/src/main/java/com/palantir/ls/api/WorkspaceCompiler.java
@@ -14,20 +14,17 @@
  * limitations under the License.
  */
 
-package com.palantir.ls.groovy;
+package com.palantir.ls.api;
 
 import io.typefox.lsapi.Diagnostic;
 import io.typefox.lsapi.FileEvent;
-import io.typefox.lsapi.ReferenceParams;
-import io.typefox.lsapi.SymbolInformation;
 import io.typefox.lsapi.TextDocumentContentChangeEvent;
 import java.net.URI;
 import java.nio.file.Path;
 import java.util.List;
-import java.util.Map;
 import java.util.Set;
 
-public interface CompilerWrapper {
+public interface WorkspaceCompiler {
 
     /**
      * Returns the root of the compiled workspace.
@@ -62,32 +59,5 @@ public interface CompilerWrapper {
      * language server.
      */
     void handleChangeWatchedFiles(List<? extends FileEvent> changes);
-
-    /**
-     * Returns a mapping from the URI of source file to symbols located within these source files.
-     */
-    Map<URI, Set<SymbolInformation>> getFileSymbols();
-
-    /**
-     * Returns a mapping from a type name (class, interface, or enum) to symbols which reference those types.
-     */
-    Map<String, Set<SymbolInformation>> getTypeReferences();
-
-    /**
-     * Returns the locations of the symbols that reference the symbol defined by the given params.
-     * @param params the parameters used to filter down which symbol is referenced
-     * @return the set of locations
-     */
-    Set<SymbolInformation> findReferences(ReferenceParams params);
-
-    /**
-     * Returns a list of symbols filtered based on a wildcard query.
-     *
-     * The character * designates zero or more of any character. The character ? designates exactly one character.
-     *
-     * @param query the query
-     * @return the set of symbols
-     */
-    Set<SymbolInformation> getFilteredSymbols(String query);
 
 }
