+++ groovy-language-server/src/main/java/com/palantir/ls/groovy/GroovyLanguageServerConfig.java
@@ -17,6 +17,7 @@
 package com.palantir.ls.groovy;
 
 import com.google.common.collect.Maps;
+import com.palantir.ls.api.CompilerWrapper;
 import io.typefox.lsapi.Diagnostic;
 import io.typefox.lsapi.MessageParams;
 import io.typefox.lsapi.PublishDiagnosticsParams;
@@ -31,11 +32,23 @@ import java.util.function.Consumer;
 
 public final class GroovyLanguageServerConfig implements LanguageServerConfig {
 
+    private CompilerWrapper compilerWrapper = null;
     private Consumer<MessageParams> showMessage = m -> { };
     private Consumer<ShowMessageRequestParams> showMessageRequest = m -> { };
     private Consumer<MessageParams> logMessage = m -> { };
     private Consumer<Object> telemetryEvent = e -> { };
     private Consumer<PublishDiagnosticsParams> publishDiagnostics = p -> { };
+
+    @Override
+    public CompilerWrapper getCompilerWrapper() {
+        return compilerWrapper;
+    }
+
+    @Override
+    public void setCompilerWrapper(CompilerWrapper compilerWrapper) {
+        this.compilerWrapper = compilerWrapper;
+    }
+
 
     @Override
     public Consumer<MessageParams> getShowMessage() {
