+++ groovy-language-server/src/test/java/com/palantir/ls/groovy/GroovycWrapperTest.java
@@ -27,6 +27,9 @@ import com.google.common.base.Optional;
 import com.google.common.collect.Iterables;
 import com.google.common.collect.Lists;
 import com.google.common.collect.Sets;
+import com.palantir.ls.groovy.compilation.GroovyTreeParser;
+import com.palantir.ls.groovy.compilation.GroovyWorkspaceCompiler;
+import com.palantir.ls.groovy.compilation.GroovycWrapper;
 import com.palantir.ls.util.DefaultDiagnosticBuilder;
 import com.palantir.ls.util.Ranges;
 import io.typefox.lsapi.Diagnostic;
@@ -73,31 +76,50 @@ public final class GroovycWrapperTest {
     @Rule
     public TemporaryFolder root = new TemporaryFolder();
 
+    private GroovycWrapper createGroovycWrapper() {
+        CompilationUnitProvider unitProvider = new DefaultCompilationUnitProvider();
+        return new GroovycWrapper(
+                GroovyWorkspaceCompiler.of(unitProvider, output.getRoot().toPath(), root.getRoot().toPath(),
+                        changedOutput.getRoot().toPath()),
+                GroovyTreeParser.of(unitProvider, root.getRoot().toPath(), changedOutput.getRoot().toPath()));
+    }
+
     @Test
     public void testTargetDirectoryNotFolder() throws IOException {
+        CompilationUnitProvider unitProvider = new DefaultCompilationUnitProvider();
         expectedException.expect(IllegalArgumentException.class);
         expectedException.expectMessage("targetDirectory must be a directory");
-        GroovycWrapper.of(output.newFile().toPath(), root.getRoot().toPath(), changedOutput.getRoot().toPath());
+        new GroovycWrapper(
+                GroovyWorkspaceCompiler.of(unitProvider, output.newFile().toPath(), root.getRoot().toPath(),
+                        changedOutput.getRoot().toPath()),
+                GroovyTreeParser.of(unitProvider, root.getRoot().toPath(), changedOutput.getRoot().toPath()));
     }
 
     @Test
     public void testWorkspaceRootNotFolder() throws IOException {
+        CompilationUnitProvider unitProvider = new DefaultCompilationUnitProvider();
         expectedException.expect(IllegalArgumentException.class);
         expectedException.expectMessage("workspaceRoot must be a directory");
-        GroovycWrapper.of(output.getRoot().toPath(), root.newFile().toPath(), changedOutput.getRoot().toPath());
+        new GroovycWrapper(
+                GroovyWorkspaceCompiler.of(unitProvider, output.getRoot().toPath(), root.newFile().toPath(),
+                        changedOutput.getRoot().toPath()),
+                GroovyTreeParser.of(unitProvider, root.newFile().toPath(), changedOutput.getRoot().toPath()));
     }
 
     @Test
     public void testChangedFilesRootNotFolder() throws IOException {
+        CompilationUnitProvider unitProvider = new DefaultCompilationUnitProvider();
         expectedException.expect(IllegalArgumentException.class);
         expectedException.expectMessage("changedFilesRoot must be a directory");
-        GroovycWrapper.of(output.getRoot().toPath(), root.getRoot().toPath(), changedOutput.newFile().toPath());
+        new GroovycWrapper(
+                GroovyWorkspaceCompiler.of(unitProvider, output.getRoot().toPath(), root.getRoot().toPath(),
+                        changedOutput.newFile().toPath()),
+                GroovyTreeParser.of(unitProvider, root.getRoot().toPath(), changedOutput.newFile().toPath()));
     }
 
     @Test
     public void testEmptyWorkspace() throws InterruptedException, ExecutionException, IOException {
-        GroovycWrapper wrapper =
-                GroovycWrapper.of(output.getRoot().toPath(), root.getRoot().toPath(), changedOutput.getRoot().toPath());
+        GroovycWrapper wrapper = createGroovycWrapper();
         Set<Diagnostic> diagnostics = wrapper.compile();
         assertEquals(0, diagnostics.size());
         Map<URI, Set<SymbolInformation>> symbols = wrapper.getFileSymbols();
@@ -140,8 +162,8 @@ public final class GroovycWrapperTest {
                         + "}\n");
         addFileToFolder(root.getRoot(), "test4.groovy", "class ExceptionNew {}");
 
-        GroovycWrapper wrapper =
-                GroovycWrapper.of(output.getRoot().toPath(), root.getRoot().toPath(), changedOutput.getRoot().toPath());
+        GroovycWrapper wrapper = createGroovycWrapper();
+
         Set<Diagnostic> diagnostics = wrapper.compile();
 
         assertEquals(0, diagnostics.size());
@@ -164,8 +186,8 @@ public final class GroovycWrapperTest {
                         + "   }\n"
                         + "}\n");
 
-        GroovycWrapper wrapper =
-                GroovycWrapper.of(output.getRoot().toPath(), root.getRoot().toPath(), changedOutput.getRoot().toPath());
+        GroovycWrapper wrapper = createGroovycWrapper();
+
         Set<Diagnostic> diagnostics = wrapper.compile();
         assertEquals(0, diagnostics.size());
         Map<URI, Set<SymbolInformation>> symbols = wrapper.getFileSymbols();
@@ -194,8 +216,8 @@ public final class GroovycWrapperTest {
                         + "   abstract double getAt(int idx);\n"
                         + "}\n");
 
-        GroovycWrapper wrapper =
-                GroovycWrapper.of(output.getRoot().toPath(), root.getRoot().toPath(), changedOutput.getRoot().toPath());
+        GroovycWrapper wrapper = createGroovycWrapper();
+
         Set<Diagnostic> diagnostics = wrapper.compile();
         assertEquals(0, diagnostics.size());
         Map<URI, Set<SymbolInformation>> symbols = wrapper.getFileSymbols();
@@ -213,8 +235,8 @@ public final class GroovycWrapperTest {
                         + "   ONE, TWO, THREE\n"
                         + "}\n");
 
-        GroovycWrapper wrapper =
-                GroovycWrapper.of(output.getRoot().toPath(), root.getRoot().toPath(), changedOutput.getRoot().toPath());
+        GroovycWrapper wrapper = createGroovycWrapper();
+
         Set<Diagnostic> diagnostics = wrapper.compile();
         assertEquals(0, diagnostics.size());
         Map<URI, Set<SymbolInformation>> symbols = wrapper.getFileSymbols();
@@ -248,8 +270,8 @@ public final class GroovycWrapperTest {
                         + "   }\n"
                         + "}\n");
 
-        GroovycWrapper wrapper =
-                GroovycWrapper.of(output.getRoot().toPath(), root.getRoot().toPath(), changedOutput.getRoot().toPath());
+        GroovycWrapper wrapper = createGroovycWrapper();
+
         Set<Diagnostic> diagnostics = wrapper.compile();
         assertEquals(0, diagnostics.size());
         Map<URI, Set<SymbolInformation>> symbols = wrapper.getFileSymbols();
@@ -282,8 +304,8 @@ public final class GroovycWrapperTest {
                         + "println name\n"
                         + "myMethod()\n");
 
-        GroovycWrapper wrapper =
-                GroovycWrapper.of(output.getRoot().toPath(), root.getRoot().toPath(), changedOutput.getRoot().toPath());
+        GroovycWrapper wrapper = createGroovycWrapper();
+
         Set<Diagnostic> diagnostics = wrapper.compile();
         assertEquals(0, diagnostics.size());
         Map<URI, Set<SymbolInformation>> symbols = wrapper.getFileSymbols();
@@ -312,8 +334,8 @@ public final class GroovycWrapperTest {
                 "interface ICoordinates {\n"
                         + "   abstract double getAt(int idx);\n"
                         + "}\n");
-        GroovycWrapper wrapper =
-                GroovycWrapper.of(output.getRoot().toPath(), root.getRoot().toPath(), changedOutput.getRoot().toPath());
+        GroovycWrapper wrapper = createGroovycWrapper();
+
         Set<Diagnostic> diagnostics = wrapper.compile();
         assertEquals(0, diagnostics.size());
 
@@ -367,8 +389,8 @@ public final class GroovycWrapperTest {
         addFileToFolder(newFolder2, "file.txt", "Something that is not groovy");
         addFileToFolder(newFolder2, "Test.java", "public class Test {}\n");
 
-        GroovycWrapper wrapper =
-                GroovycWrapper.of(output.getRoot().toPath(), root.getRoot().toPath(), changedOutput.getRoot().toPath());
+        GroovycWrapper wrapper = createGroovycWrapper();
+
         Set<Diagnostic> diagnostics = wrapper.compile();
 
         assertEquals(0, diagnostics.size());
@@ -410,8 +432,8 @@ public final class GroovycWrapperTest {
                         + "}\n");
         addFileToFolder(root.getRoot(), "test4.groovy", "class ExceptionNew {}\n");
 
-        GroovycWrapper wrapper =
-                GroovycWrapper.of(output.getRoot().toPath(), root.getRoot().toPath(), changedOutput.getRoot().toPath());
+        GroovycWrapper wrapper = createGroovycWrapper();
+
         Set<Diagnostic> diagnostics = wrapper.compile();
 
         assertEquals(2, diagnostics.size());
@@ -453,9 +475,8 @@ public final class GroovycWrapperTest {
                         + "   class InnerCat2 {\n"
                         + "   }\n"
                         + "}\n");
+        GroovycWrapper wrapper = createGroovycWrapper();
 
-        GroovycWrapper wrapper =
-                GroovycWrapper.of(output.getRoot().toPath(), root.getRoot().toPath(), changedOutput.getRoot().toPath());
         Set<Diagnostic> diagnostics = wrapper.compile();
         assertEquals(0, diagnostics.size());
 
@@ -485,8 +506,8 @@ public final class GroovycWrapperTest {
         // edge case on one line
         File enumFile = addFileToFolder(newFolder1, "MyEnum.groovy",
                 "enum MyEnum {ONE,TWO}\n");
-        GroovycWrapper wrapper =
-                GroovycWrapper.of(output.getRoot().toPath(), root.getRoot().toPath(), changedOutput.getRoot().toPath());
+        GroovycWrapper wrapper = createGroovycWrapper();
+
         Set<Diagnostic> diagnostics = wrapper.compile();
         assertEquals(0, diagnostics.size());
 
@@ -530,8 +551,8 @@ public final class GroovycWrapperTest {
                         + "A a\n"
                         + "B b\n"
                         + "}\n");
-        GroovycWrapper wrapper =
-                GroovycWrapper.of(output.getRoot().toPath(), root.getRoot().toPath(), changedOutput.getRoot().toPath());
+        GroovycWrapper wrapper = createGroovycWrapper();
+
         Set<Diagnostic> diagnostics = wrapper.compile();
         assertEquals(0, diagnostics.size());
 
@@ -612,8 +633,8 @@ public final class GroovycWrapperTest {
                 "abstract class AbstractCoordinates {\n"
                         + "   abstract void something();\n"
                         + "}\n");
-        GroovycWrapper wrapper =
-                GroovycWrapper.of(output.getRoot().toPath(), root.getRoot().toPath(), changedOutput.getRoot().toPath());
+        GroovycWrapper wrapper = createGroovycWrapper();
+
         Set<Diagnostic> diagnostics = wrapper.compile();
         assertEquals(0, diagnostics.size());
         Map<String, Set<SymbolInformation>> references = wrapper.getTypeReferences();
@@ -695,8 +716,8 @@ public final class GroovycWrapperTest {
                 "class Cat {\n"
                         + "   public String name = \"Bobby\"\n"
                         + "}\n");
-        GroovycWrapper wrapper =
-                GroovycWrapper.of(output.getRoot().toPath(), root.getRoot().toPath(), changedOutput.getRoot().toPath());
+        GroovycWrapper wrapper = createGroovycWrapper();
+
         Set<Diagnostic> diagnostics = wrapper.compile();
         assertEquals(0, diagnostics.size());
 
@@ -749,8 +770,8 @@ public final class GroovycWrapperTest {
         File catFile = addFileToFolder(newFolder1, "Cat.groovy",
                 "class Cat {\n"
                         + "}\n");
-        GroovycWrapper wrapper =
-                GroovycWrapper.of(output.getRoot().toPath(), root.getRoot().toPath(), changedOutput.getRoot().toPath());
+        GroovycWrapper wrapper = createGroovycWrapper();
+
         Set<Diagnostic> diagnostics = wrapper.compile();
         assertEquals(0, diagnostics.size());
 
@@ -792,8 +813,8 @@ public final class GroovycWrapperTest {
                 "enum Animal {\n"
                         + "CAT, DOG, BUNNY\n"
                         + "}\n");
-        GroovycWrapper wrapper =
-                GroovycWrapper.of(output.getRoot().toPath(), root.getRoot().toPath(), changedOutput.getRoot().toPath());
+        GroovycWrapper wrapper = createGroovycWrapper();
+
         Set<Diagnostic> diagnostics = wrapper.compile();
         assertEquals(0, diagnostics.size());
         Set<SymbolInformation> expectedResult = Sets.newHashSet(
@@ -862,8 +883,8 @@ public final class GroovycWrapperTest {
         File catFile = addFileToFolder(newFolder1, "Cat.groovy",
                 "class Cat {\n"
                         + "}\n");
-        GroovycWrapper wrapper =
-                GroovycWrapper.of(output.getRoot().toPath(), root.getRoot().toPath(), changedOutput.getRoot().toPath());
+        GroovycWrapper wrapper = createGroovycWrapper();
+
         Set<Diagnostic> diagnostics = wrapper.compile();
         assertEquals(0, diagnostics.size());
 
@@ -897,8 +918,8 @@ public final class GroovycWrapperTest {
         File catFile = addFileToFolder(newFolder1, "Cat.groovy",
                 "class Cat {\n"
                         + "}\n");
-        GroovycWrapper wrapper =
-                GroovycWrapper.of(output.getRoot().toPath(), root.getRoot().toPath(), changedOutput.getRoot().toPath());
+        GroovycWrapper wrapper = createGroovycWrapper();
+
         // Compile
         assertEquals(0, wrapper.compile().size());
         Set<SymbolInformation> symbols = wrapper.getFileSymbols().get(catFile.toURI());
@@ -951,8 +972,8 @@ public final class GroovycWrapperTest {
         File catFile = addFileToFolder(newFolder1, "Cat.groovy",
                 "class Cat {\n"
                         + "}\n");
-        GroovycWrapper wrapper =
-                GroovycWrapper.of(output.getRoot().toPath(), root.getRoot().toPath(), changedOutput.getRoot().toPath());
+        GroovycWrapper wrapper = createGroovycWrapper();
+
         // Compile
         assertEquals(0, wrapper.compile().size());
         Set<SymbolInformation> symbols = wrapper.getFileSymbols().get(catFile.toURI());
@@ -1003,8 +1024,8 @@ public final class GroovycWrapperTest {
         File catFile = addFileToFolder(newFolder1, "Cat.groovy",
                 "class Cat {\n"
                         + "}\n");
-        GroovycWrapper wrapper =
-                GroovycWrapper.of(output.getRoot().toPath(), root.getRoot().toPath(), changedOutput.getRoot().toPath());
+        GroovycWrapper wrapper = createGroovycWrapper();
+
         // Compile
         assertEquals(0, wrapper.compile().size());
         Set<SymbolInformation> symbols = wrapper.getFileSymbols().get(catFile.toURI());
@@ -1056,8 +1077,8 @@ public final class GroovycWrapperTest {
         File catFile = addFileToFolder(newFolder1, "Cat.groovy",
                 "class Cat {\n"
                         + "}\n");
-        GroovycWrapper wrapper =
-                GroovycWrapper.of(output.getRoot().toPath(), root.getRoot().toPath(), changedOutput.getRoot().toPath());
+        GroovycWrapper wrapper = createGroovycWrapper();
+
         // Compile
         assertEquals(0, wrapper.compile().size());
         Set<SymbolInformation> symbols = wrapper.getFileSymbols().get(catFile.toURI());
@@ -1110,8 +1131,8 @@ public final class GroovycWrapperTest {
         File catFile = addFileToFolder(newFolder1, "Cat.groovy",
                 "class Cat {\n"
                         + "}\n");
-        GroovycWrapper wrapper =
-                GroovycWrapper.of(output.getRoot().toPath(), root.getRoot().toPath(), changedOutput.getRoot().toPath());
+        GroovycWrapper wrapper = createGroovycWrapper();
+
         // Compile
         assertEquals(0, wrapper.compile().size());
         Set<SymbolInformation> symbols = wrapper.getFileSymbols().get(catFile.toURI());
@@ -1141,8 +1162,8 @@ public final class GroovycWrapperTest {
     @Test
     public void testHandleChangeWatchedFiles_created() throws IOException {
         File newFolder1 = root.newFolder();
-        GroovycWrapper wrapper =
-                GroovycWrapper.of(output.getRoot().toPath(), root.getRoot().toPath(), changedOutput.getRoot().toPath());
+        GroovycWrapper wrapper = createGroovycWrapper();
+
         // Compile
         assertEquals(0, wrapper.compile().size());
         // Assert no symbols exist
